requirements.txt
streamlit==1.26.0
Pillow==9.2.0
groq==0.3.0